Load the current term from the head block on startup
Fixes failure to sync when restarting the network that would happen
after block production was stopped and restarted through the admin API
or other conditions that would advance the term. The new blocks used to
have their term set to 1, which meant that even though they were
correctly transmitted and received, nodes other than the producer would
consider them worse than the current head and fail to switch to them.