Updated compactor to optimize memory layout
binary search nodes now organize their keys based on the probability
that a particular key will be compared in a lower_bound search and
prefetch is used to load the top of the tree. This optimization is
performed by the compactor when it moves because it slows down inserts
if it does it at that time.

Improved jitter in measurement of lowerbound/sec in multi-thread test