Fix bug that caused gc_queue to run cleanup too early
The code assumed that the _sequence for a session is in bounds, but this
was not necesarily true. If the session lock was suspended right after
reading _end, the state of the queue could change arbitrarily before it
writes _sequence. session::lock accounts for this, but it does create a
transient out-of-bounds value for _sequence that may be seen by
start_wait. This would cause us to remove too many items from the queue.
Clearing these elements was fine, because they were already nullptr.
Decrementing _size, though, is a big problem, because it is how we know
which items in the queue are free and safe to reuse.