Fix uninitialized variable
reset is called from accept_block_header, which is run on all nodes. The
timer isn't necessarily running. The extra calls to reset are harmless.
We might end up lowering the timeout while being a non-producer, but
that seems okay, since we shouldn't receive blocks any slower when we
are a block producer.