Remove special case for no hosts
Allowing extra root hosts is potentially a security risk if it allows
accessing x-admin in an unexpected way and bypasses authentication. The
only thing that we actually need this for is loading local services,
which we handle using a special host "\0". Since null is not a
field-vchar, it is impossible for a client to submit a request with it
in the host (Even with a malicious client, psinode will fail to parse
such malformed requests).