Fix race when sending HTTP requests
The socket isn't known until after the call to sendRequest returns, so
the service has to insert it in a table afterwards. However, the
response might come back at any time after sendRequest, and it could be
before the table entry was added, causing the response to get lost.

The solution is to add some extra rules:
- The socket for an outgoing request starts with autoClose enabled.
- recv cannot be run while autoClose is enabled. If a response is
  received, it will be delayed until autoClose is disabled.
- setting autoClose will fail, if recv is running in a different
  context.

The interface in XHttp is refactored to require a separate call to
setCallback instead of passing the callbacks to setRequest.

Provided the service inserts the socket in a table and calls setCallback
in a single subjective tx, this ensures that:
- If the service aborts before inserting the socket into a table,
  everything related to the socket will be cleaned up correctly.
- When recv is called, it will see the table entry